refactor module structure, linting
