+++ api/qupo_backend/config.py
@@ -10,7 +10,7 @@ class Settings(BaseSettings):
     azure_name: str = ''
     azure_subscription_id: str = ''
     ibmq_client_secret: str = ''
-    nasdaq_api_key: str = 'rN2sRJT1wqfut9CA-PUy'
+    nasdaq_api_key: str = ''
     sqllite_db_url: str = 'sqlite:///./app/db/finance.db'
 
     class Config:
