+++ api/test/opti_backend_runscript.py
@@ -1,7 +1,5 @@
-# native packages
 import sys
 import os.path
-# custom packages
 import qupo_backend.optimization_backend.opti_backend_runner as obr
 import stockdata_integrator_runscript as sdi
 import qupo_backend.finance_utilities as fu
