+++ docs/202205_QuPO_Arch_v01.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2022-05-31T09:39:29.438Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/18.1.3 Chrome/100.0.4896.143 Electron/18.2.3 Safari/537.36" etag="1OxkyRg5CN560HE-sC5A" version="18.1.3" type="device" pages="2"><diagram id="_edhJc90cHTKVlBYo0jf" name="Page-1">7R1bc9o6+tcwk+5MGMtXeAyQXs7knKale3a7LzsCBLgxNrVNEvrrj2RZxpZkbINtnDRth2Ih3767vpt62njz/MGH2/Wf3gI5PVVZPPe0SU9VVVMZ4P/IyJ6ODIYWHVj59oIOgcPA1P6F4kElHt3ZCxRkJoae54T2Njs491wXzcPMGPR97yk7bek52btu4QoJA9M5dMTR/9iLcB2/hWodxj8ie7VmdwbmkP6ygWxy/CbBGi68p9SQdtvTxr7nhfTb5nmMHAI8Bhd63vucX5MH85EbljlhPbZvr9Hdj5/Pv/5YX99c//VlrV4nj/sInV38yl9vp9/wyM0CbkPkxw8f7hlE8HtsydeNt9iRsRF07JWLBxy0xA8yCrZwbruru+hogh/qMGOOHxVfURs9Ij+0MZBv4h9Cb4tH4yfBv6Hn3JcECegwzSFvg0J/j6cwghvE0I7JjQH/6YA7zYzH1im8JRNhTC+r5NIHkOIvMVSrQFizBAgLMF353m6b+/oxbcMZm65UBgsHFUMESzKWBstg2BxYBgJY3vsepg930TWSO4ZXEeQxjK9VrSSQAWiQ+BQByj6+so0eUQRib/7QU8fkfT0/xMLR9qLHCFAQbMj7qybcEDi5s4CCSzh2CAJmWEyYK/INEvQdrhYQSUiVAodTDHAyfx1u8BtNAMFi6HsPaOw5HsbXxPVcguml7TjcUCFqN/ZiEZHJ09oO0RQTB7nnE1ZQeAxzmrtABGhKM/iXs5wmYTmggr5qSIRRY+RgaAI5LNDSdlGM9ZDgK5h7WzJwhfqrPqMS8oNjB5giFIwLfB8MxGf8vws36N1vh1xOy2A12gnsGgJ2t773iA0oPOjbwQO5a8Sft9MPBAw+WiIMiTk2sH43DA6zGNSHXUCgKbJnSlr7KNhhafu7YYrjNWvQBUxpIqu1b9SVgIylSgyOxAKuHS4So+7zNrQ39i8Y2p6LfxnB+cPLMfEMOQIuB+CBKQD46w4TkmSZ9mRvHEhZH5vV0/gXUB5w87XtLO7g3tuR1wpCjDl2NFp7vv0LXxYeZA70w3j5rumZGVNyZiwmsAzDc+4Z5AE39Cd8zky8g0EYD8w9x4HbwJ4lr7GB/sp2R14YepsapFAOruPrGGpWX4ispoOBTAIN8ikhvtlXIgbclYMOdwMKx9tAkaxkEwGQuaOZvSF0MI5dGKIRkdeBQIHJq55BlLIV7ojIOAcblT3tBn+99z2Muc3VfXplwNxFk/fvoln40TVA8cidPvUcjNWre2rO+HTNcuOsMJGF6018eP8Jf449d2mv3h272NdIkV4dnfOHN+sZo/ip6fXpM9Dv9BI9Y5Knjqvq4FjMyTkxK/z0w8jXGO9kyMPnLp3Iw7TGehu5kY4OIVMx5DW3nu2GEWEYI/wPk89Y6RvRa4zxMTgc439kuh9ieOJ3gXbEPQjz4xMiPHkGq1EZlrtwxtKjFGuZZj5nnSdjRSUmYNmxI+xRLDPHIDgJxYmRFeP0G0H55BoIeNdEvGsSHDtwhpx7L7AjjatNfDqXw/3F0AuUkqLziOA8D72ix5Ny/xhGJ8+YhYLXRzH3Y9N76fmbnUOtmLHau1EE4UEFz84nYjzxkxwuO3a83SJtAClT5D/aZO0lvx7WiEGx2f8mZ04nxEE5QkwotnZSZIvOFCkylYjBEb1ztww75TUadpyjzhSoAsjsrGOe8ddi2A1Bjqhc2ojE0YihFO6Js/BNOFUky6Eip0pGhRInRxmiqE80qW9GUHPoNUoKmaZsoKHoaKSMjS+/9hZXtrvdhXRNpmC9gA/eWLxuGjDFdU67LA4YFaawihYrxCwKDJS1t/Jc6NweRjm374/dZsvmQ39O0JOcdedFTEx4+AcKw31sRMBd6GX91ejZDv9LvvcVRY2Pv5Mb9BXDiI8nzH6IDvapg3vk2xggxN5JmQrkRU4yFDA4vJ0/RyVWENgyWqGjFJDjwPURWUk8Zh+wAfxa3cKvksHtccw2j0UqA89ATnzqPZE1BwtPt0DWwOMNRUo18VkHFN/4PtynpsUiLPc+BuAMSaClL1c8f6hwFEafoFYDEhiiL2WFXMyxczF1gE8uSJxvcs3zggNex1ckPAXpQ0FR6KZETzQXVxY9JgcnSbBF8/4bMnOXl8DKsh0LJFwMmyw21RWtwL53QytQ3V67VjAZEyd5iRx+6YMJWkG4kGXyYpxbI9SkXnSLT6RUjj+XfnR+Q+rFFB0Uc+JujRYuAYnX9NIO3VhWHRy7v5swGvJIkqxBVK0ZaeT+Wv/fdvcm/Lj5/OHL387/nr/9fS1JWcL2fEATzUgMjmSa4Y/JqNPIKi+ABLTmp3dwnK5JAidJfmdLqkMM9Poo3Plu7xVnJ1VkKV2MKjTGUvIcHNFam5IMzgUMIeEk4iiBUWTpakKH7uAe+WISZ7WUnHbz/E2DA7skvVazJMyRcEztcGdPkAP3TxgkKx+GUfbs1WgX2C4Kgrqgf+kyCxn4ZXUWySq0dk0CRP90VbO2JdcGhrG/TyxfckDPUg12fDgvOspavwKWslatHDg9wWElV8c9mYF9rjVc1fa0WPCBaT51eNz25DVldv7ZtucxkPLcLrdVGDPP9w5JnPdzlFaKsGaUMu9myQC2YlcRvX7ehTTyEo3nEkV5S4OHtzoQPQ5JEDwTnqghPiEF7vmO6VN4mfFlzMAJZzbGllpJttTrZsuzkCPGjqYkR0BAWRYhBfReAxlrXCRfkqMJZNm6dSTrSgF1vvv9HI2UaKHvqV8KNFJrlK9fSiGdhVDRmb2MzDkE5+teUk4nwfkdCUdn8VR+eRNlzOQGNydSPByjx9LFCLmcdq30uSAHPTrVScameMtlgEIBTTVoa12UWSiqZVsyvIklUJeWX0CSGdSuAGPr1jcJliuZCiUYQ1hHRBiQeEzXKLJb7YQX8P/xqtUmK1T4CG0nkg0vzxtXmQtNS8twocqM44wxPBS5sDEfKRANrghVe1K2SYoRVpEo4zC39L3NK0eVzhXlaIy406gyG6oLlKNKLM+KUOV6aUx9h2uP1NkvbRe6cyTg7tXiS8/GQjVNE/ClytaZjeGL+UrfFJxEcUk0nNxFrqp1qzi5n0fj/BQWH0TNiX3WZUSqkmCVtyVVIYSHpZVtn2c/SFeii5uWWhZypiqLUrRqWr5xXj5srLK25cWWx/IIyCVRWiWN5IDSkzzwR/NPzkGyXLy2hFL5ulm0PT+Q/D1Iz0SQCLsk1yupecNY3GEZQmtpJalh3Mi/8CljsgqZeaSnTIAcLDRpv4DMAl1yJ9tdRDV00mtOadrHjyiPgHY+If00kB3lhdDhTxNqmv3coRiuJI4dLXuI6iHCnbh4QnsThUxhsHfn7+S3i/vjLOxl1FaFfF/u3OhNsCUW2nmP+RdGlvuA77Qjd7pmjzY+GIZch6bMYarFEhn343YA5DvMFkijcH6eJpLHE7nAZA3q6tpQOFeuuAbThhJ1pRtN6StDXDfT9BQWVm816CMY9jUAHShcTB0AWc8qIMllGDQVU5esp94ni6Z/hzbjKh7ylyuRVM1el0skC3wfXCA1idYX1qg0ViIL8npLpEoQo9Vuzqr5rT6pgLlEYmCOr5I1as3VJ731YWgQvZI89FZLEMGRNgxztqa+STpRXqX6I9wRO8sYTalVaEziQsVPUb/KQxOFwq/83diiHctjGFWUBZK7fsOG4BT5RO1o77mbv4mgemlU1lVRRqNAb0r9qKLZJ6C56T5vfFzQkMUFmYMnDRZLbcgYZg8gd4SxPKhmfF/H0VRs7V7O4yXasi8GaCzZOUOHEvugXYCKpuFLAyhLWt1nD9sAqdyDJ+ZJtC/uuEJVaTairO4BNJZbLNECHyQxV743dgsklyDsfME31PvqwLCATj+1pmCpvgJYSuWh1U+BD38OjHLc3BLgxeDVi4U7X8pu9NNwt5jPoH3AS9+PFSl2IC5yUlREay8oUjap6iJlCXzNkRZXT+aVJQjzY29moyWxpsjmNPDx5NuRXRT7q0OSgLKYCZTZpWSTWn2ZsmSTxGJopTDPknk8WjaugKLx1pW4mJRtUmI1JR2tEtvjpKmR7aJEyGQBg3VCM+XpLymte16RTav6dJ8otU/aGnxyb8gRuTiBkNkHxKEf7WcFtD4YkhuRCR8RJDdWjb4xOMsFYuU0HEj3h5BFuhqj0hJe1+7jY2H7cQgZS5hzvVR5KJJbIxI50yoCGUNfVsyUKBAFsghuc/W5ugiXdH1ust2ZcoUV2vVi1rmS3CPYPr7aaxfOgw6lAamnVeJabeYBlS4kyGlE0LDRy5lRpn68zZgw39I4imqiD4wiWr03v6JW3EX7lX359FkkV8ext0FaM82j5t2FC96ctIyz7FgA+I4/iiWRpbK06TqSYeSt8yTbvEy/3HdNYFbMdxnI9oBqFa7J3dIu7f3nrei+flmQHcq0f7uQNbvS6zRROX1F0bJqZ2ipBXonOjqh4amk/0Ob3c+EVSdQmNor6H9WVV0NrJwb5emrohOaUlhiROGLHTzY5MlSOQfZzbNetAzAr3xxIaCoOWYCvwOHzCp4aeC+uJWAHyEH3MVW2Q664U5s0Nd1y+zyJC6p7xWg2PTqX3C9Sgwr6RYXjWVvi3zPajCTdOK6KE0EVgGiKkS9SqZDNUhdon26X+ZAsIvekhjqx90lLTOsSJsXZ1hdFwup5QzbVClZsjlTCix/wWABf/bY7sUfd7PCcPWZXHwaianluDjbOuVyPM16GaQg/ROr34W4ZXs3OToH3JfjaMZ/lyptjSIkqVLIFrfQyEPG2etHPmjZH5rA0AY6/TS44GROHbh4XXyhgXL4wyoOktYf2cvmrFFrWxQmPuPue847RyF87yS9ryTkgT9Z54nKFMJLa71dkgAX7cBUQ031EYcWSKIsp3u0TskoOmbath1e0bm2+VqslnJ3ceFziqrOH7bh3WJc0oFevLk0Cxg588SXlwZnNRITPL7O6mYiXAFRnU20Qv/f7PymiHbwG5htpaPSjWhpvlGzDqy+njHkSmnp+lB+0SY0dRtcFaVL/djlcn50Re0baSOsZexqXQn5td50qB4yYY6o4qhhThVns1pI6AZmFZhCzKkln9+QVmFArNAzjOzKsSQVyAL5nlB9cVaEg28bZsniG7LKisaSG4F6GdMyxyYsyhNLEmJBJd4s1xa7ATdLBttmyT3KqjIusOT3yV2DW7z5qLbBuLItknmH6AMK52tGVrRDz5h64+NB0n6G1Un0VA3/fU8eYrTy4cJGh98kZRVsOqm+l9ZgZLKtmV/1jus6MYtbxBT0bkr8symij1rnUI4Aai/bSify7ZLt1MiLLu1n8hy0pQDybx8R7SwAepLc7qdA7++CumKwSbo6q41ne1WmJBSbkilnaKythXqZ9NcTBZRo4NQgshppY5eTvnOayBKuo3IJSkZDu/bqA0v6wM3KslImSK/OWvmzmJoHkmG22K1UbuF2yA3brZz2YwTXjkyontGeXQcYg6KM9ux8E7SR0S5pbrGB4RrhD6LFCeO+/E1MC1YfhkWoNgN7WSaRovcb6vGegxoxNWFGt/P7LbDCRSxMyZqwfZRcxpHXJaOrwZqhg0g818wSzB+2LXbNZhYLQJ24wqw83xi0oBKYMyZd2r/d4oFoOfVqBQ5Jlu/zradkneHUw7R2pI725iJ4ES4Cs69lxY6lieTTrpdAExttxXXIwZwgPeok/mo5ejjsm1nP8kCT5EG2ztCGyL8dNiNONBpKBH7K7kuq1WtdPGfxyxuYVY0NjfPpaCbIdOhqxvTgTAPDLMh1yE5vJSZlXDiGUi2k2SqZNxjfFB0Lfd1I/R2aWcKxaiJ7vjl/TXSe7J2VmPLHCZ331Fab3s3gLlNi7TJShiFSXEWqMVJ8RQqMzQLeeqH1xMnuUQkxcem9ZXnFMLJkcyhUrptbeK9J/MR5DybML0iZ47tuVJueVI/mrts5A5o/oSn+Ol9R1Z3Bo+opZzxmMSwFq6bpnWrv1aj7WEvd4kSDZsJ2JqvEZvQEBpw1VpaDB8OiK9XEwYKoYDfKlS08k1U/YVigIy3+3SufUPBIkj2BsmdUZXx86HtemJ5O1vw09Um7/Qc=</diagram><diagram id="UB9cr3BmkRtWqjOSIXLo" name="Processes">tZRBT4MwFIB/TY9LoHWIRzanXtQlzM14q/QNmhSKXTeYv95iCwwxMUvmifZr+8r73gNE5nl9r2iZPUoGAmGP1YjcIoyxFwTm0ZCjJb4fEEtSxZljPYj5JzjoObrnDHaDjVpKoXk5hIksCkj0gFGlZDXctpVieGtJUxiBOKFiTDec6czSEF/3/AF4mukuvxu7ktN2s8tkl1EmqxNEFojMlZTajvJ6DqKx13qJtmoZKc9bTz6ip7f16mXzqic22N05R7oUFBT6sqGxDX2gYu98LeoSFM+bm2zW+tiqVHJfMGjCeYjMqoxriEuaNKuV6R7DMp0LM/PNkAqeFmYsYGveeXYApbmpSuSwls1+d7tZg/pHuf7I1e8KYFoXZA5aHc05FyV0JXNNO3XTqm+AK+JYdlL8sIXUNV3aRe7FmoFze4ZnMvK8fI5XCAc0bzwU77tex4WM/5dbTKYDueFYro9/kRuc79ZM+8/re+3kL0UWXw==</diagram></mxfile>
